requirements.txt
streamlit>=1.36
requests>=2.31
beautifulsoup4>=4.12
openai-whisper
torch
ffmpeg-python
